Extract runtime ResourcePage generation modules
Move generated page orchestration and generated-page upsert behavior out of the
runtime weave facade into page_generation.ts. Move the remaining ResourcePage
model assembly helpers into page_model_assembly.ts, keeping timing phase names
stable and preserving executeGenerate/executeWeave behavior.

Update the runtime codebase overview with the new module layout.
